% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{EAP}
\alias{EAP}
\alias{EAP,}
\alias{Rasch}
\alias{Method}
\title{Finding expected a posteriori value for theta}
\usage{
EAP(raschObj, lower = -6, upper = 6)
}
\arguments{
\item{lower}{A numeric object for lower bound of integration}

\item{upper}{A numeric object for upper bound of integration}

\item{Raschobj}{A Rasch object}
}
\value{
An object of class Numeric that represents the EAP value for theta
}
\description{
Finds EAP of theta using \code{Likelihood} and \code{Prior} functions
}
\note{
This function is for test takers
}
\examples{

myRasch <- new("Rasch", "Tyler", c(1,2,3), c(1,0,1)) 
myLower <- -6
myUpper <- 6
EAP(myRasch, myLower, myUpper)
}
\seealso{
\code{\link{Likelihood}}, \code{\link{Prior}}
}
\author{
Tyler Brandt
}
