% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{Likelihood}
\alias{Likelihood}
\alias{Likelihood,}
\alias{Rasch}
\alias{Method}
\title{Finding likelihood that theta estimate is accurate}
\usage{
Likelihood(raschObj, theta)
}
\arguments{
\item{theta}{A numeric object.}

\item{Raschobj}{A Rasch object}
}
\value{
An object of class Numeric that represents the likelihood that the estimated theta is accurate
}
\description{
Finds product of PQ vector
}
\note{
This function is for test takers
}
\examples{

myRasch <- new("Rasch", "Tyler", c(1,2,3), c(1,0,1)) 
myTheta <- 3
Likelihood(myRasch, myTheta)
}
\seealso{
\code{\link{Probability}}
}
\author{
Tyler Brandt
}
