% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior}
\alias{Prior}
\alias{Prior,}
\alias{Numeric}
\alias{Method}
\title{Finding prior for a given value of theta}
\usage{
Prior(theta)
}
\arguments{
\item{theta}{A numeric object.}
}
\value{
An object of class Numeric that represents a prior for theta
}
\description{
Finds height of normal curve with mean = 0, sd = 3 at theta
}
\note{
This function is for test takers
}
\examples{

myTheta <- 3
Prior(myTheta)
}
\author{
Tyler Brandt
}
