% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\alias{Rasch}
\alias{Initialize,}
\alias{Validitiy,}
\alias{Print}
\alias{print,Rasch-method}
\title{A student who took a test with his difficulty and answers}
\usage{
\S4method{print}{Rasch}(x)
}
\description{
Object of class \code{Rasch} are used in the \code{Probability}, \code{Likelihood}, \code{Prior}, \code{EAP} functions
}
\details{
An object of the class 'Rasch' has the following slots:
\itemize{
\item \code{name} The name of the student who took the test
\item \code{a} The difficulties of the questions on the test
\item \code{y_j} The student's answers
}
}
\author{
Tyler Brandt" \email{brandttyler@wustl.edu}
}
